fixed several classes and tests to make everything run
git-svn-id: http://svn.newfolderconsulting.com/prt@293 0c11b213-6b92-4ae5-b10c-e53e59125080

+++ data/prtDataProstate.m
@@ -34,7 +34,7 @@ cellData{3} = cat(1,cellData{3}{:});
 targetInd = 5;
 featureInd = setdiff(2:length(cellData),targetInd);
 Y = cellData{targetInd};
-X = cat(2,cellData{featureInd});  %drop the counting var, and the Y
+X = double(cat(2,cellData{featureInd}));  %drop the counting var, and the Y
 colNames = varNames(featureInd);
 
 varNames = cellfun(@(c)c{1},varNames,'uniformoutput',false);
